Move dtos folders
