+++ src/LoreWeave.Api/Dtos/Maps/KnowsDtoMapper.cs
@@ -1,3 +1,4 @@
+using LoreWeave.Api.Dtos.Knows;
 using LoreWeave.Application.Models;
 
 namespace LoreWeave.Api.Dtos.Maps;
